+++ src/main/java/com/rezztoran/rezztoranbe/controller/RestaurantTableController.java
@@ -33,9 +33,7 @@ public class RestaurantTableController {
   @GetMapping("/{id}")
   public ResponseEntity<ApiResponse<Object>> getTableById(@PathVariable Long id) {
     var restaurantTable = tableService.getTableById(id);
-    RestaurantTableDTO restaurantTableDTO =
-        modelMapper.map(restaurantTable, RestaurantTableDTO.class);
-    restaurantTableDTO.setRestaurantId(restaurantTable.getRestaurant().getId());
+    RestaurantTableDTO restaurantTableDTO = convertDto(restaurantTable);
     return ApiResponse.builder().data(restaurantTableDTO).build();
   }
 
@@ -46,50 +44,44 @@ public class RestaurantTableController {
     if (localDate != null) {
       List<RestaurantTable> restaurantTables =
           tableService.getAvailableTablesByRestaurantAndDate(id, localDate);
-      return getApiResponseResponseEntity(restaurantTables);
+      return convertDto(restaurantTables);
     }
     List<RestaurantTable> restaurantTables = tableService.getTablesByRestaurant(id);
-    return getApiResponseResponseEntity(restaurantTables);
-  }
-
-  private ResponseEntity<ApiResponse<Object>> getApiResponseResponseEntity(
-      List<RestaurantTable> restaurantTables) {
-    var restaurantTableDtos =
-        restaurantTables.stream()
-            .map(
-                restaurantTable -> {
-                  RestaurantTableDTO restaurantTableDTO =
-                      modelMapper.map(restaurantTable, RestaurantTableDTO.class);
-                  restaurantTableDTO.setRestaurantId(restaurantTable.getRestaurant().getId());
-                  return restaurantTableDTO;
-                })
-            .collect(Collectors.toList());
-    return ApiResponse.builder().data(restaurantTableDtos).build();
+    return convertDto(restaurantTables);
   }
 
   @PostMapping
   public ResponseEntity<ApiResponse<Object>> createTable(
       @RequestBody RestaurantTableRequestModel requestModel) {
     var restaurantTable = tableService.createTable(requestModel);
-    RestaurantTableDTO restaurantTableDTO =
-        modelMapper.map(restaurantTable, RestaurantTableDTO.class);
-    restaurantTableDTO.setRestaurantId(restaurantTable.getRestaurant().getId());
+    RestaurantTableDTO restaurantTableDTO = convertDto(restaurantTable);
     return ApiResponse.builder().data(restaurantTableDTO).build();
   }
 
-  @PutMapping("/{id}")
+  @PutMapping
   public ResponseEntity<ApiResponse<Object>> updateTableById(
       @RequestBody RestaurantTableRequestModel requestModel) {
     var restaurantTable = tableService.updateTable(requestModel);
+    RestaurantTableDTO restaurantTableDTO = convertDto(restaurantTable);
+    return ApiResponse.builder().data(restaurantTableDTO).build();
+  }
+
+  private RestaurantTableDTO convertDto(RestaurantTable restaurantTable) {
     RestaurantTableDTO restaurantTableDTO =
         modelMapper.map(restaurantTable, RestaurantTableDTO.class);
     restaurantTableDTO.setRestaurantId(restaurantTable.getRestaurant().getId());
-    return ApiResponse.builder().data(restaurantTableDTO).build();
+    return restaurantTableDTO;
   }
 
   @DeleteMapping("/{id}")
   public ResponseEntity<ApiResponse<Object>> deleteTableById(@PathVariable Long id) {
     tableService.deleteTableById(id);
     return ApiResponse.builder().build();
+  }
+
+  private ResponseEntity<ApiResponse<Object>> convertDto(List<RestaurantTable> restaurantTables) {
+    var restaurantTableDtos =
+        restaurantTables.stream().map(this::convertDto).collect(Collectors.toList());
+    return ApiResponse.builder().data(restaurantTableDtos).build();
   }
 }
